refactored custom object action
