+++ packages/destination-actions/src/destinations/hubspot/upsertObject/constants.ts
@@ -12,3 +12,5 @@ export const SUPPORTED_HUBSPOT_OBJECT_TYPES = [
 ]
 
 export const MAX_HUBSPOT_BATCH_SIZE = 100
+
+export const OBJECT_NOT_FOUND_ERROR_RESPONSE = 'Unable to infer object type'
